Vehicle Parts refactored
